[TECH-404] Move deployset logic to discovery
branch: feature/TECH-404-link-up-deploy-set
